Extract game functions to helper, add logger
