+++ .gitignore
@@ -2,3 +2,5 @@ node_modules
 .cache
 dist
 coverage
+combined.log
+error.log
