+++ server/package.json
@@ -15,7 +15,8 @@
     "passport-steam": "^1.0.8",
     "passport.socketio": "^3.7.0",
     "socket.io": "^2.0.4",
-    "socket.io-client": "^2.0.4"
+    "socket.io-client": "^2.0.4",
+    "winston": "^3.0.0-rc1"
   },
   "devDependencies": {
     "babel-cli": "^6.26.0",
