+++ server/src/socket/game.js
@@ -1,4 +1,4 @@
-import _ from 'lodash'
+import logger from '../logger'
 import db from '../db'
 import actions from '../../../common/actions.json'
 
@@ -23,32 +23,3 @@ export async function getGameRequest(io, socket) {
     socket.emit('action', { type: actions.GET_GAME_ERROR })
   }
 }
-
-export async function shouldStartCaptainVote(gameId) {
-  const game = await db('game')
-    .first('*')
-    .where({ id: gameId })
-  const numberOfPlayers = await db('player')
-    .count('id')
-    .where({ gameId })
-  return parseInt(numberOfPlayers[0].count, 10) === game.maxPlayers
-}
-
-export async function startCaptainVote(gameId) {
-  const games = await db('game')
-    .update({ status: 'voting captains' })
-    .where({ id: gameId })
-    .returning('*')
-  return _.first(games)
-}
-
-export function createNewGame() {
-  return db('game').insert({})
-}
-
-export function getGameStatus(gameId) {
-  return db('game')
-    .first('status')
-    .where({ id: gameId })
-}
-
